Added new controller to show collections.
