+++ quadriga/src/main/java/edu/asu/diging/quadriga/web/AddCollectionController.java
@@ -18,40 +18,31 @@ import edu.asu.diging.quadriga.web.forms.CollectionForm;
 @Controller
 public class AddCollectionController {
 
-    @Autowired
-    private ICollectionManager collectionManager;
-    
-    @Autowired
-    private CollectionRepository collectionRepo;
+	@Autowired
+	private ICollectionManager collectionManager;
 
-    @RequestMapping(value = "/auth/collections/add", method = RequestMethod.GET)
-    public String get(Model model) {
-        model.addAttribute("collectionForm", new CollectionForm());
-        return "admin/user/addCollection";
-    }
-    
-	
-    @RequestMapping(value="/auth/collections",method=RequestMethod.GET) 
-    public String showCollection(Model model) {
-    	model.addAttribute("collections",collectionRepo.findAll()); 
-    	return "admin/user/showcollection"; 
-    }
-	 
 
-    @RequestMapping(value = "/auth/collections/add", method = RequestMethod.POST)
-    public String add(@Valid CollectionForm collectionForm, BindingResult result,
-            RedirectAttributes redirectAttrs) {
-        if (result.hasErrors()) {
-            return "admin/user/addCollection";
-        }
-        Collection collection=new Collection();
-        collection.setName(collectionForm.getName());
-        collection.setDescription(collection.getDescription());
-        collectionManager.addCollection(collection);
-        
-        redirectAttrs.addFlashAttribute("alert_type", "success");
-        redirectAttrs.addFlashAttribute("alert_msg", "Collection has been added.");
-        return "redirect:/auth/collections";
-    }
+
+	@RequestMapping(value = "/auth/collections/add", method = RequestMethod.GET)
+	public String get(Model model) {
+		model.addAttribute("collectionForm", new CollectionForm());
+		return "admin/user/addCollection";
+	}
+
+	@RequestMapping(value = "/auth/collections/add", method = RequestMethod.POST)
+	public String add(@Valid CollectionForm collectionForm, BindingResult result,
+			RedirectAttributes redirectAttrs) {
+		if (result.hasErrors()) {
+			return "admin/user/addCollection";
+		}
+		Collection collection=new Collection();
+		collection.setName(collectionForm.getName());
+		collection.setDescription(collection.getDescription());
+		collectionManager.addCollection(collection);
+
+		redirectAttrs.addFlashAttribute("alert_type", "success");
+		redirectAttrs.addFlashAttribute("alert_msg", "Collection has been added.");
+		return "redirect:/auth/collections";
+	}
 
 }
